[metasrv] refactor: remove dup method get_tables()
